Relocating ChunkedUploadProvider - thx @daboxu
